get-issue-tests + moving pojo class
