+++ src/main/java/jira/api/issue/getissueresponse/Description.java
@@ -1,5 +1,5 @@
 
-package jira.api.issue;
+package jira.api.issue.getissueresponse;
 
 import java.util.List;
 import javax.annotation.Generated;
