+++ src/main/java/jira/api/issue/getissueresponse/Status.java
@@ -1,16 +1,19 @@
 
-package jira.api.issue;
+package jira.api.issue.getissueresponse;
 
 import javax.annotation.Generated;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 @Generated("jsonschema2pojo")
-public class Priority {
+public class Status {
 
     @SerializedName("self")
     @Expose
     private String self;
+    @SerializedName("description")
+    @Expose
+    private String description;
     @SerializedName("iconUrl")
     @Expose
     private String iconUrl;
@@ -20,6 +23,9 @@ public class Priority {
     @SerializedName("id")
     @Expose
     private String id;
+    @SerializedName("statusCategory")
+    @Expose
+    private StatusCategory statusCategory;
 
     public String getSelf() {
         return self;
@@ -27,6 +33,14 @@ public class Priority {
 
     public void setSelf(String self) {
         this.self = self;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
     }
 
     public String getIconUrl() {
@@ -51,6 +65,14 @@ public class Priority {
 
     public void setId(String id) {
         this.id = id;
+    }
+
+    public StatusCategory getStatusCategory() {
+        return statusCategory;
+    }
+
+    public void setStatusCategory(StatusCategory statusCategory) {
+        this.statusCategory = statusCategory;
     }
 
 }
