+++ src/main/java/jira/api/issue/getissueresponse/Worklog.java
@@ -1,5 +1,5 @@
 
-package jira.api.issue;
+package jira.api.issue.getissueresponse;
 
 import java.util.List;
 import javax.annotation.Generated;
@@ -7,38 +7,27 @@ import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 @Generated("jsonschema2pojo")
-public class Comment {
+public class Worklog {
 
-    @SerializedName("comments")
+    @SerializedName("startAt")
     @Expose
-    private List<Object> comments;
-    @SerializedName("self")
-    @Expose
-    private String self;
+    private Integer startAt;
     @SerializedName("maxResults")
     @Expose
     private Integer maxResults;
     @SerializedName("total")
     @Expose
     private Integer total;
-    @SerializedName("startAt")
+    @SerializedName("worklogs")
     @Expose
-    private Integer startAt;
+    private List<Object> worklogs;
 
-    public List<Object> getComments() {
-        return comments;
+    public Integer getStartAt() {
+        return startAt;
     }
 
-    public void setComments(List<Object> comments) {
-        this.comments = comments;
-    }
-
-    public String getSelf() {
-        return self;
-    }
-
-    public void setSelf(String self) {
-        this.self = self;
+    public void setStartAt(Integer startAt) {
+        this.startAt = startAt;
     }
 
     public Integer getMaxResults() {
@@ -57,12 +46,12 @@ public class Comment {
         this.total = total;
     }
 
-    public Integer getStartAt() {
-        return startAt;
+    public List<Object> getWorklogs() {
+        return worklogs;
     }
 
-    public void setStartAt(Integer startAt) {
-        this.startAt = startAt;
+    public void setWorklogs(List<Object> worklogs) {
+        this.worklogs = worklogs;
     }
 
 }
